telegram-bot/bot.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-

"""
Simple Telegram Bot placeholder for the Snel Pointless project.
This is a basic implementation that will be expanded later.
"""

import os
import logging
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Set up logging
logging.basicConfig(
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    level=logging.INFO
)

logger = logging.getLogger(__name__)

def main():
    """Start the bot."""
    # This is a placeholder for the actual bot implementation
    logger.info("Telegram bot started")
    logger.info("This is a placeholder implementation")
    logger.info("The actual bot will be implemented in a future update")
    
    # Keep the script running
    try:
        while True:
            pass
    except KeyboardInterrupt:
        logger.info("Bot stopped by user")

if __name__ == '__main__':
    main()
